# backend-express/swagger.yaml
openapi: 3.0.0
info:
  title: Kopi Haus Dashboard API
  version: 1.0.0
  description: API documentation for the Kopi Haus Dashboard backend application.
servers:
  - url: http://localhost:5001/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRegister:
      type: object
      required:
        - name
        - email
        - password
        - roles
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: password123
        roles:
          type: array
          items:
            type: string
            enum: [Cashier, Warehouse, Admin, Manager]
          example: [Cashier]
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered successfully
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Server error
...
    Register:
      type: object
      required:
        - outlet_id
        - name
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c4
        outlet_id:
          type: string
          example: 60d5ec49f8c7a40015a7b8c1
        name:
          type: string
          example: Main Register
        status:
          type: string
          enum: [idle, open, closed]
          example: idle
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T10:00:00.000Z'
...
    Return:
      type: object
      required:
        - type
        - outlet_id
        - processed_by
        - returned_at
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c20
        type:
          type: string
          enum: [sales, purchase]
          description: Type of return (sales or purchase)
          example: sales
        ref_sale_id:
          type: string
          nullable: true
          description: Reference to the Sale ID if type is 'sales'
          example: 60d5ec49f8c7a40015a7b8c17
        ref_purchase_id:
          type: string
          nullable: true
          description: Reference to the Purchase ID if type is 'purchase'
          example: 60d5ec49f8c7a40015a7b8c15
        outlet_id:
          type: string
          description: ID of the outlet where the return was processed
          example: 60d5ec49f8c7a40015a7b8c1
        processed_by:
          type: string
          description: ID of the user who processed the return
          example: 60d5ec49f8c7a40015a7b8c13
        returned_at:
          type: string
          format: date-time
          description: Timestamp when the return was processed
          example: '2023-10-26T11:00:00.000Z'
        reason:
          type: string
          nullable: true
          description: Reason for the return
          example: Damaged item
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T11:00:00.000Z'
    ReturnItem:
      type: object
      required:
        - return_id
        - product_id
        - qty
        - amount
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c21
        return_id:
          type: string
          description: ID of the associated return
          example: 60d5ec49f8c7a40015a7b8c20
        product_id:
          type: string
          description: ID of the returned product
          example: 60d5ec49f8c7a40015a7b8c6
        qty:
          type: number
          format: integer
          description: Quantity of the product returned
          example: 1
        amount:
          type: number
          format: float
          description: Amount of the returned product (price per unit)
          example: 25000
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T11:00:00.000Z'
...
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: User already exists or invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /registers:
    get:
      summary: Get all registers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of registers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Register'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new register
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - outlet_id
                - name
              properties:
                outlet_id:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c1
                name:
                  type: string
                  example: New Register Name
                status:
                  type: string
                  enum: [idle, open, closed]
                  example: idle
      responses:
        '201':
          description: Register created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Register'
        '400':
          description: Invalid input or Outlet ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /registers/{id}:
    get:
      summary: Get a register by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the register to retrieve
      responses:
        '200':
          description: Register data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Register'
        '404':
          description: Register not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update a register by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the register to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                outlet_id:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c1
                name:
                  type: string
                  example: Updated Register Name
                status:
                  type: string
                  enum: [idle, open, closed]
                  example: open
      responses:
        '200':
          description: Register updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Register'
        '400':
          description: Invalid input or Outlet ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Register not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a register by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the register to delete
      responses:
        '200':
          description: Register deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Register not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /returns:
    get:
      summary: Get all returns
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of returns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Return'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new return with items
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - outlet_id
                - processed_by
                - items
              properties:
                type:
                  type: string
                  enum: [sales, purchase]
                  example: sales
                ref_sale_id:
                  type: string
                  nullable: true
                  example: 60d5ec49f8c7a40015a7b8c17
                ref_purchase_id:
                  type: string
                  nullable: true
                  example: 60d5ec49f8c7a40015a7b8c15
                outlet_id:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c1
                processed_by:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c13
                reason:
                  type: string
                  nullable: true
                  example: Customer changed mind
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - product_id
                      - qty
                      - amount
                    properties:
                      product_id:
                        type: string
                        example: 60d5ec49f8c7a40015a7b8c6
                      qty:
                        type: number
                        format: integer
                        example: 1
                      amount:
                        type: number
                        format: float
                        example: 25000
      responses:
        '201':
          description: Return created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  return:
                    $ref: '#/components/schemas/Return'
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReturnItem'
        '400':
          description: Invalid input, reference ID, outlet ID, user ID, or product ID, or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /returns/{id}:
    get:
      summary: Get a return by ID with its items
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the return to retrieve
      responses:
        '200':
          description: Return data with items
          content:
            application/json:
              schema:
                type: object
                properties:
                  return:
                    $ref: '#/components/schemas/Return'
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReturnItem'
        '404':
          description: Return not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a return by ID (and its items)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the return to delete
      responses:
        '200':
          description: Return and its items deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Return not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...